fix tests and rename
