+++ backend/src/handlers/actions/PokerTableJoinHandler.test.ts
