+++ backend/src/handlers/actions/PokerTableJoinHandler.ts
@@ -1,51 +1,53 @@
 // External modules
-import { Request, Response } from 'express';
-import Rooms from '../../sockets/Rooms';
-import GameLobbyService from '../../game-lobby-service';
+import { Response } from 'express';
 
 // Internal modules
 import BaseHandler from '../../shared/BaseHandler';
 import Result from '../../shared/Result';
 import Logger from '../../utils/Logger';
+import Rooms from '../../sockets/Rooms';
+import GameLobbyService from '../../game-lobby-service';
 
 // Types
 import {
-  PokerTableLeavePayload,
-  PokerTableLeaveOutput,
-  pokerTableLeaveSchema,
-} from '../../shared/api/types/PokerTableLeave';
+  PokerTableJoinPayload,
+  PokerTableJoinOutput,
+  pokerTableJoinSchema,
+} from '../../shared/api/types/PokerTableJoin';
 
 const debug = Logger.newDebugger('APP:Routes:actions');
 
-class TablesLeaveHandler extends BaseHandler<PokerTableLeavePayload, PokerTableLeaveOutput> {
+class PokerTableJoinHandler extends BaseHandler<PokerTableJoinPayload, PokerTableJoinOutput> {
   constructor() {
-    super(pokerTableLeaveSchema);
+    super(pokerTableJoinSchema);
   }
 
-  protected getResult(payload: Result<PokerTableLeavePayload>, res: Response<PokerTableLeaveOutput>) {
+  protected getResult(payload: Result<PokerTableJoinPayload>, res: Response<PokerTableJoinOutput>) {
     const seatNumber = payload.getValue().selectedSeatNumber;
     const clientId = payload.getValue().socketId;
+
     const pokerTable = GameLobbyService.getTable('table_1');
+
     if (!pokerTable) {
       return res.send({
         ok: false,
         error: 'Table does not exist',
       });
     }
-    const leave_room = pokerTable.leaveTable(seatNumber, clientId);
-    if (!leave_room.ok) {
+    const join_room = pokerTable.sitAtTable('table_1', seatNumber, clientId);
+    if (!join_room.ok) {
       return res.send({
         ok: false,
-        error: leave_room.errorMessage,
+        error: join_room.errorMessage,
       });
     }
     // Emit event to all clients connected that a player has sat down
-    const event = 'player_left';
+    const event = 'player_joined';
     const eventPayload = {
       playerId: clientId,
       seatId: seatNumber,
     };
-    const send_events = Rooms.sendEventToRoom('table_1', event, payload);
+    const send_events = Rooms.sendEventToRoom('table_1', event, eventPayload);
     if (!send_events.ok) {
       return res.send({
         ok: false,
@@ -56,4 +58,4 @@ class TablesLeaveHandler extends BaseHandler<PokerTableLeavePayload, PokerTableL
   }
 }
 
-export default TablesLeaveHandler;
+export default PokerTableJoinHandler;
