+++ backend/src/handlers/actions/PokerTableLeaveHandler.ts
@@ -1,53 +1,51 @@
 // External modules
-import { Request, Response } from 'express';
+import { Response } from 'express';
+import Rooms from '../../sockets/Rooms';
+import GameLobbyService from '../../game-lobby-service';
 
 // Internal modules
 import BaseHandler from '../../shared/BaseHandler';
 import Result from '../../shared/Result';
 import Logger from '../../utils/Logger';
-import Rooms from '../../sockets/Rooms';
-import GameLobbyService from '../../game-lobby-service';
 
 // Types
 import {
-  PokerTableJoinPayload,
-  PokerTableJoinOutput,
-  pokerTableJoinSchema,
-} from '../../shared/api/types/PokerTableJoin';
+  PokerTableLeavePayload,
+  PokerTableLeaveOutput,
+  pokerTableLeaveSchema,
+} from '../../shared/api/types/PokerTableLeave';
 
 const debug = Logger.newDebugger('APP:Routes:actions');
 
-class TablesJoinHandler extends BaseHandler<PokerTableJoinPayload, PokerTableJoinOutput> {
+class PokerTableLeaveHandler extends BaseHandler<PokerTableLeavePayload, PokerTableLeaveOutput> {
   constructor() {
-    super(pokerTableJoinSchema);
+    super(pokerTableLeaveSchema);
   }
 
-  protected getResult(payload: Result<PokerTableJoinPayload>, res: Response<PokerTableJoinOutput>) {
+  protected getResult(payload: Result<PokerTableLeavePayload>, res: Response<PokerTableLeaveOutput>) {
     const seatNumber = payload.getValue().selectedSeatNumber;
     const clientId = payload.getValue().socketId;
-
     const pokerTable = GameLobbyService.getTable('table_1');
-
     if (!pokerTable) {
       return res.send({
         ok: false,
         error: 'Table does not exist',
       });
     }
-    const join_room = pokerTable.sitAtTable('table_1', seatNumber, clientId);
-    if (!join_room.ok) {
+    const leave_room = pokerTable.leaveTable(seatNumber, clientId);
+    if (!leave_room.ok) {
       return res.send({
         ok: false,
-        error: join_room.errorMessage,
+        error: leave_room.errorMessage,
       });
     }
     // Emit event to all clients connected that a player has sat down
-    let event = 'player_joined';
-    let eventPayload = {
+    const event = 'player_left';
+    const eventPayload = {
       playerId: clientId,
       seatId: seatNumber,
     };
-    let send_events = Rooms.sendEventToRoom('table_1', event, eventPayload);
+    const send_events = Rooms.sendEventToRoom('table_1', event, eventPayload);
     if (!send_events.ok) {
       return res.send({
         ok: false,
@@ -58,4 +56,4 @@ class TablesJoinHandler extends BaseHandler<PokerTableJoinPayload, PokerTableJoi
   }
 }
 
-export default TablesJoinHandler;
+export default PokerTableLeaveHandler;
