+++ backend/src/routes/apiMethods.ts
@@ -5,6 +5,6 @@ export interface ApiMethodShape {
 }
 export const ApiMethods: { [key: string]: ApiMethodShape } = {
   // add entries for each API method
-  tablesJoin: { httpMethod: 'post', path: 'tables.join', handler: '../handlers/actions/TablesJoinHandler.ts' },
-  tablesLeave: { httpMethod: 'post', path: 'tables.leave', handler: '../handlers/actions/TablesLeaveHandler.ts' },
+  tablesJoin: { httpMethod: 'post', path: 'tables.join', handler: '../handlers/actions/PokerTableJoinHandler.ts' },
+  tablesLeave: { httpMethod: 'post', path: 'tables.leave', handler: '../handlers/actions/PokerTableLeaveHandler.ts' },
 } as const;
